requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
cymem==2.0.11
dataclasses==0.6
deepface==0.0.93
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djongo==1.2.31
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mongoengine==0.29.1
mtcnn==1.0.0
murmurhash==1.0.13
namex==0.1.0
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
preshed==3.0.10
protobuf==5.29.5
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pymongo==4.13.0
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.2
pywin32==310
redis==6.2.0
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
spotipy==2.25.1
sqlparse==0.5.3
srsly==2.5.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
